# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Beneficial Ownership Change Request Workflow System API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  success: boolean;
  message: string;
}

export interface LoginInput {
  email: string;
  /** @minLength 1 */
  password: string;
}

export type UserRole = typeof UserRole[keyof typeof UserRole];


export const UserRole = {
  APPLICANT: 'APPLICANT',
  REVIEWER: 'REVIEWER',
} as const;

export interface User {
  id: number;
  email: string;
  role: UserRole;
}

export interface AuthResponse {
  token: string;
  user: User;
}

export type ApplicationStatus = typeof ApplicationStatus[keyof typeof ApplicationStatus];


export const ApplicationStatus = {
  DRAFT: 'DRAFT',
  SUBMITTED: 'SUBMITTED',
  UNDER_REVIEW: 'UNDER_REVIEW',
  APPROVED: 'APPROVED',
  REJECTED: 'REJECTED',
  CHANGES_REQUESTED: 'CHANGES_REQUESTED',
} as const;

export interface Application {
  id: number;
  companyName: string;
  registrationNumber: string;
  beneficialOwnerName: string;
  ownershipPercentage: number;
  changeReason: string;
  /** @nullable */
  supportingNotes?: string | null;
  status: ApplicationStatus;
  userId: number;
  createdAt: string;
  updatedAt: string;
  user?: User;
}

export interface AuditLogEntry {
  id: number;
  applicationId: number;
  userId: number;
  userRole: UserRole;
  previousStatus: ApplicationStatus;
  newStatus: ApplicationStatus;
  /** @nullable */
  comment?: string | null;
  timestamp: string;
  user?: User;
}

export interface ApplicationDetail {
  id: number;
  companyName: string;
  registrationNumber: string;
  beneficialOwnerName: string;
  ownershipPercentage: number;
  changeReason: string;
  /** @nullable */
  supportingNotes?: string | null;
  status: ApplicationStatus;
  userId: number;
  createdAt: string;
  updatedAt: string;
  user?: User;
  auditLog: AuditLogEntry[];
}

export interface ApplicationInput {
  /** @minLength 1 */
  companyName: string;
  /** @minLength 1 */
  registrationNumber: string;
  /** @minLength 1 */
  beneficialOwnerName: string;
  /**
     * @minimum 0
     * @maximum 100
     */
  ownershipPercentage: number;
  /** @minLength 1 */
  changeReason: string;
  supportingNotes?: string;
}

export interface ApplicationUpdate {
  /** @minLength 1 */
  companyName?: string;
  /** @minLength 1 */
  registrationNumber?: string;
  /** @minLength 1 */
  beneficialOwnerName?: string;
  /**
     * @minimum 0
     * @maximum 100
     */
  ownershipPercentage?: number;
  /** @minLength 1 */
  changeReason?: string;
  supportingNotes?: string;
}

export interface WorkflowActionInput {
  comment?: string;
}

export interface WorkflowActionWithComment {
  /** @minLength 1 */
  comment: string;
}

export type ApplicationStatsByStatus = {[key: string]: number};

export interface ApplicationStats {
  total: number;
  byStatus: ApplicationStatsByStatus;
}

export type ListApplicationsParams = {
status?: ApplicationStatus;
};