# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Beneficial Ownership Change Request Workflow System API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: auth
    description: Authentication
  - name: applications
    description: Ownership change request applications

paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current authenticated user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current user info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /applications:
    get:
      operationId: listApplications
      tags: [applications]
      summary: List applications (filtered by role)
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/ApplicationStatus"
      responses:
        "200":
          description: List of applications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Application"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: createApplication
      tags: [applications]
      summary: Create a new ownership change request (applicant only)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApplicationInput"
      responses:
        "201":
          description: Application created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Application"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /applications/{id}:
    get:
      operationId: getApplication
      tags: [applications]
      summary: Get a single application with audit log
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Application with audit history
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApplicationDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateApplication
      tags: [applications]
      summary: Update a DRAFT application (owner only)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApplicationUpdate"
      responses:
        "200":
          description: Updated application
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Application"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /applications/{id}/submit:
    post:
      operationId: submitApplication
      tags: [applications]
      summary: Submit a DRAFT application for review (applicant only)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Application submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Application"
        "400":
          description: Illegal workflow transition
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /applications/{id}/start-review:
    post:
      operationId: startReview
      tags: [applications]
      summary: Move application to UNDER_REVIEW (reviewer only)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Review started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Application"
        "400":
          description: Illegal workflow transition
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /applications/{id}/approve:
    post:
      operationId: approveApplication
      tags: [applications]
      summary: Approve an UNDER_REVIEW application (reviewer only)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WorkflowActionInput"
      responses:
        "200":
          description: Application approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Application"
        "400":
          description: Illegal workflow transition
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /applications/{id}/reject:
    post:
      operationId: rejectApplication
      tags: [applications]
      summary: Reject an UNDER_REVIEW application (reviewer only, comment required)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WorkflowActionWithComment"
      responses:
        "200":
          description: Application rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Application"
        "400":
          description: Illegal workflow transition or missing comment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /applications/{id}/request-changes:
    post:
      operationId: requestChanges
      tags: [applications]
      summary: Request changes on an UNDER_REVIEW application (reviewer only, comment required)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WorkflowActionWithComment"
      responses:
        "200":
          description: Changes requested
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Application"
        "400":
          description: Illegal workflow transition or missing comment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /applications/stats:
    get:
      operationId: getApplicationStats
      tags: [applications]
      summary: Get application counts grouped by status
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Application stats by status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApplicationStats"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    ErrorResponse:
      type: object
      required: [success, message]
      properties:
        success:
          type: boolean
        message:
          type: string

    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 1

    AuthResponse:
      type: object
      required: [token, user]
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"

    User:
      type: object
      required: [id, email, role]
      properties:
        id:
          type: integer
        email:
          type: string
        role:
          $ref: "#/components/schemas/UserRole"

    UserRole:
      type: string
      enum: [APPLICANT, REVIEWER]

    ApplicationStatus:
      type: string
      enum:
        - DRAFT
        - SUBMITTED
        - UNDER_REVIEW
        - APPROVED
        - REJECTED
        - CHANGES_REQUESTED

    Application:
      type: object
      required:
        - id
        - companyName
        - registrationNumber
        - beneficialOwnerName
        - ownershipPercentage
        - changeReason
        - status
        - userId
        - createdAt
        - updatedAt
      properties:
        id:
          type: integer
        companyName:
          type: string
        registrationNumber:
          type: string
        beneficialOwnerName:
          type: string
        ownershipPercentage:
          type: number
        changeReason:
          type: string
        supportingNotes:
          type: ["string", "null"]
        status:
          $ref: "#/components/schemas/ApplicationStatus"
        userId:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        user:
          $ref: "#/components/schemas/User"

    ApplicationDetail:
      type: object
      required:
        - id
        - companyName
        - registrationNumber
        - beneficialOwnerName
        - ownershipPercentage
        - changeReason
        - status
        - userId
        - createdAt
        - updatedAt
        - auditLog
      properties:
        id:
          type: integer
        companyName:
          type: string
        registrationNumber:
          type: string
        beneficialOwnerName:
          type: string
        ownershipPercentage:
          type: number
        changeReason:
          type: string
        supportingNotes:
          type: ["string", "null"]
        status:
          $ref: "#/components/schemas/ApplicationStatus"
        userId:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        user:
          $ref: "#/components/schemas/User"
        auditLog:
          type: array
          items:
            $ref: "#/components/schemas/AuditLogEntry"

    ApplicationInput:
      type: object
      required:
        - companyName
        - registrationNumber
        - beneficialOwnerName
        - ownershipPercentage
        - changeReason
      properties:
        companyName:
          type: string
          minLength: 1
        registrationNumber:
          type: string
          minLength: 1
        beneficialOwnerName:
          type: string
          minLength: 1
        ownershipPercentage:
          type: number
          minimum: 0
          maximum: 100
        changeReason:
          type: string
          minLength: 1
        supportingNotes:
          type: string

    ApplicationUpdate:
      type: object
      properties:
        companyName:
          type: string
          minLength: 1
        registrationNumber:
          type: string
          minLength: 1
        beneficialOwnerName:
          type: string
          minLength: 1
        ownershipPercentage:
          type: number
          minimum: 0
          maximum: 100
        changeReason:
          type: string
          minLength: 1
        supportingNotes:
          type: string

    WorkflowActionInput:
      type: object
      properties:
        comment:
          type: string

    WorkflowActionWithComment:
      type: object
      required: [comment]
      properties:
        comment:
          type: string
          minLength: 1

    AuditLogEntry:
      type: object
      required:
        - id
        - applicationId
        - userId
        - userRole
        - previousStatus
        - newStatus
        - timestamp
      properties:
        id:
          type: integer
        applicationId:
          type: integer
        userId:
          type: integer
        userRole:
          $ref: "#/components/schemas/UserRole"
        previousStatus:
          $ref: "#/components/schemas/ApplicationStatus"
        newStatus:
          $ref: "#/components/schemas/ApplicationStatus"
        comment:
          type: ["string", "null"]
        timestamp:
          type: string
          format: date-time
        user:
          $ref: "#/components/schemas/User"

    ApplicationStats:
      type: object
      required: [total, byStatus]
      properties:
        total:
          type: integer
        byStatus:
          type: object
          additionalProperties:
            type: integer