# lib/api-zod/src/generated/types/applicationDetail.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Beneficial Ownership Change Request Workflow System API
 * OpenAPI spec version: 0.1.0
 */
import type { ApplicationStatus } from './applicationStatus';
import type { AuditLogEntry } from './auditLogEntry';
import type { User } from './user';

export interface ApplicationDetail {
  id: number;
  companyName: string;
  registrationNumber: string;
  beneficialOwnerName: string;
  ownershipPercentage: number;
  changeReason: string;
  /** @nullable */
  supportingNotes?: string | null;
  status: ApplicationStatus;
  userId: number;
  createdAt: Date;
  updatedAt: Date;
  user?: User;
  auditLog: AuditLogEntry[];
}