# lib/api-zod/src/generated/types/applicationInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Beneficial Ownership Change Request Workflow System API
 * OpenAPI spec version: 0.1.0
 */

export interface ApplicationInput {
  /** @minLength 1 */
  companyName: string;
  /** @minLength 1 */
  registrationNumber: string;
  /** @minLength 1 */
  beneficialOwnerName: string;
  /**
     * @minimum 0
     * @maximum 100
     */
  ownershipPercentage: number;
  /** @minLength 1 */
  changeReason: string;
  supportingNotes?: string;
}