# lib/api-zod/src/generated/types/applicationStats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Beneficial Ownership Change Request Workflow System API
 * OpenAPI spec version: 0.1.0
 */
import type { ApplicationStatsByStatus } from './applicationStatsByStatus';

export interface ApplicationStats {
  total: number;
  byStatus: ApplicationStatsByStatus;
}